requirements.txt
flask
loguru
pandas
playwright
requests
schedule
yfinance
python-dotenv
exchange-calendars
gunicorn
flask-compress
